Add option to include prebuilt images when signing target files.
When producing a signed build using sign_target_files_apks.py and a prebuilt
vendor.img (or other prebuilt images that are to be included in the final zip)
sign_target_files_apks.py must be told to include the prebuilt images or they
will not be included in the final file.

The build process uses the make configuration statements
BOARD_PREBUILT_VENDORIMAGE and AB_OTA_PARTITIONS to inform the build process
to add the image files, however since sign_target_files_apks.py is not part
of the make system, it does not know about these configuration settings.

This patch therefore adds a new command line option to the
sign_target_files_apks.py script to allow the user to add one or more images
to the signed-target_files.zip that is produced.

The command line option has the following format:

  --prebuilts_path <path to prebuild image files>
      Specify a path that contains one or more image files, to be added to the
      signed_target-files.zip.  Note, this directory must contain *only* .img
      files that you want to add. All other files in the directory will be
      added as well.

Change-Id: I34fcc8b4f94b3494458c3be376571ca26f15335e