Regenerate the hashtree for prebuilt vendor.img files.
When building with a prebuilt vendor.img file (via BOARD_PREBUILT_VENDORIMAGE)
the vendor.img may not have the same hashtree and salt as the ones being
created during the build process.

To make sure that vendor.img matches system/boot/dtbo, the hashtree should be
rewritten using avbtool.

This patch therefor adds in the rewrite of the hashtree if using a prebuilt
vendor.img and AVB is enabled.  The code is similar to the code in the
Add_Dtbo() function in this same script.

Change-Id: I7754d6bb5c9c7719d81e48e9d202c5fae17200c0